Rename App.tsx to app.tsx.
